+++ src/components/callsToAction/ActionButtons.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { StyleSheet, css } from 'aphrodite';
 
-const styles = {
+const styles = StyleSheet.create({
     buttonsWrapper: {
         marginTop: '2rem',
         marginBottom: '0rem',
@@ -69,36 +69,34 @@ const styles = {
         height: '1.5rem',
         paddingTop: '0.5rem'
     }
-};
-
-const style = StyleSheet.create(styles)
+});
 
 const twitterMessage = 'https://matchus.us/aclu-matchers/ just helped my donation catch fire. Drag, drop and multiply your contribution, too! #lightamatch #aclumatch';
 
 const ActionButtons = () => {
     return (
         <div>
-            <div className={ css(style.buttonsWrapper)}>
-                <div className={ css(style.buttonWrapper)}>
-                    <h2 className={css(style.caption)}>Haven't donated yet?</h2>
-                    <a className={ css(style.button) } href="https://action.aclu.org/secure/donate-to-aclu" target="_blank">
-                        <div className={ css(style.label) }>Donate to the American Civil Liberties Union</div>
+            <div className={ css(styles.buttonsWrapper)}>
+                <div className={ css(styles.buttonWrapper)}>
+                    <h2 className={css(styles.caption)}>Haven't donated yet?</h2>
+                    <a className={ css(styles.button) } href="https://action.aclu.org/secure/donate-to-aclu" target="_blank">
+                        <div className={ css(styles.label) }>Donate to the American Civil Liberties Union</div>
                     </a>
                 </div>
                 
-                <div className={ css(style.buttonWrapper)}>
-                    <h2 className={css(style.caption)}>Dropped your receipt?</h2>
-                    <a className={ css(style.button) } href={`https://twitter.com/intent/tweet?text=${twitterMessage}`}>
-                        <div className={ css(style.label) }>Spread the fire on Twitter!</div>
-                        <svg className={css(style.icon)} viewBox="0 0 1792 1792" xmlns="http://www.w3.org/2000/svg"><path d="M1684 408q-67 98-162 167 1 14 1 42 0 130-38 259.5t-115.5 248.5-184.5 210.5-258 146-323 54.5q-271 0-496-145 35 4 78 4 225 0 401-138-105-2-188-64.5t-114-159.5q33 5 61 5 43 0 85-11-112-23-185.5-111.5t-73.5-205.5v-4q68 38 146 41-66-44-105-115t-39-154q0-88 44-163 121 149 294.5 238.5t371.5 99.5q-8-38-8-74 0-134 94.5-228.5t228.5-94.5q140 0 236 102 109-21 205-78-37 115-142 178 93-10 186-50z" fill="#fff"/></svg>
+                <div className={ css(styles.buttonWrapper)}>
+                    <h2 className={css(styles.caption)}>Dropped your receipt?</h2>
+                    <a className={ css(styles.button) } href={`https://twitter.com/intent/tweet?text=${twitterMessage}`}>
+                        <div className={ css(styles.label) }>Spread the fire on Twitter!</div>
+                        <svg className={css(styles.icon)} viewBox="0 0 1792 1792" xmlns="http://www.w3.org/2000/svg"><path d="M1684 408q-67 98-162 167 1 14 1 42 0 130-38 259.5t-115.5 248.5-184.5 210.5-258 146-323 54.5q-271 0-496-145 35 4 78 4 225 0 401-138-105-2-188-64.5t-114-159.5q33 5 61 5 43 0 85-11-112-23-185.5-111.5t-73.5-205.5v-4q68 38 146 41-66-44-105-115t-39-154q0-88 44-163 121 149 294.5 238.5t371.5 99.5q-8-38-8-74 0-134 94.5-228.5t228.5-94.5q140 0 236 102 109-21 205-78-37 115-142 178 93-10 186-50z" fill="#fff"/></svg>
                     </a>
                 </div>
             </div>
 
-            <div className={ css(style.callToAction) }>
+            <div className={ css(styles.callToAction) }>
                 <div>You can make a big difference by encouraging others to use this site.</div>
                 <div>Please help us spread this spark!</div>
-                <img className={ css(style.lightAMatch) } src='./lightamatch.png' />
+                <img className={ css(styles.lightAMatch) } src='./lightamatch.png' />
             </div>
         </div>
     );
